fix: rename the file is_loading.dart to loading.dart
